envoy: extract the endpoint restoration wait into one function
Both xDS server implementations open with the same twenty lines: wait for the
restorer under a deadline, then wait for the restored endpoints to have computed
their initial policy, return on cancellation and log a warning on expiry. Move
it to one function so the next change has a single place to alter.

No behaviour change. The existing tests for the success, deadline, cancelled and
nil-promise paths pass untouched.

AIL:3
