+++ pkg/envoy/grpcnew.go
@@ -44,33 +44,19 @@ func (s *adsServer) startAdsGRPCServer(ctx context.Context) error {
 
 	reflection.Register(grpcServer)
 
-	restoreCtx, cancel := context.WithTimeout(ctx, s.config.policyRestoreTimeout)
-	defer cancel()
 	s.stopFunc = grpcServer.Stop
 
-	if s.restorerPromise != nil {
-		s.logger.Info("Envoy: Waiting for endpoint restorer before serving xDS resources...")
-		restorer, err := s.restorerPromise.Await(restoreCtx)
-		if err == nil && restorer != nil {
-			s.logger.Info("Envoy: Waiting for endpoint restoration before serving xDS resources...")
-			err = restorer.WaitForInitialPolicy(restoreCtx)
-		}
-		if errors.Is(err, context.Canceled) {
-			s.logger.Debug("Envoy: xDS server stopped before started serving")
-			return err
-		}
-		if errors.Is(err, context.DeadlineExceeded) {
-			s.logger.Warn("Envoy: Endpoint policy restoration took longer than configured restore timeout, starting serving resources to Envoy",
-				logfields.Duration, s.config.policyRestoreTimeout,
-			)
-		}
+	if err := awaitEndpointPolicyRestoration(ctx, s.logger, s.restorerPromise,
+		s.config.policyRestoreTimeout); err != nil {
+		s.logger.Debug("Envoy: xDS server stopped before started serving")
+		return err
 	}
 
 	s.logger.Info("Envoy: Starting xDS gRPC server listening",
 		logfields.Address, listener.Addr(),
 	)
 
-	ctx, cancel = context.WithCancel(ctx)
+	ctx, cancel := context.WithCancel(ctx)
 	defer cancel()
 	go func() {
 		<-ctx.Done()
